% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_matrix.R
\name{G_matrix}
\alias{G_matrix}
\title{Generate the G matrix for variance estimation in low dimension scenario.}
\usage{
G_matrix(z, u, z.sigma_est)
}
\arguments{
\item{z}{Vector. The fully observed observed shadow variable.}

\item{u}{Matrix. The covariate matrix excluding shadow variable \eqn{z} (fully observed).}

\item{z.sigma_est}{The estimated standard error of \eqn{z}.}
}
\value{
\eqn{G} matrix for variance estimation in low dimension scenario.
}
\description{
Generate the G matrix for variance estimation in low dimension scenario.
}
