% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIsMiss_se.R
\name{NFMiss_se}
\alias{NFMiss_se}
\title{Estimate standard error using asymptotic formula for unregularized regression.}
\usage{
NFMiss_se(y, z, u, gamma_tilde_hat, beta_hat, se.method, M = 500, seed_num,
  regularize, data_pair, newcov_pair, obs_id, w_hat)
}
\arguments{
\item{y}{Vector. The outcome variable}

\item{z}{Vector. The shadow variable (fully observed).}

\item{u}{Matrix. The covariate matrix \eqn{U}.}

\item{gamma_tilde_hat}{Initial estimates of scaled \eqn{\gamma}.}

\item{beta_hat}{Initial estimates of \eqn{\beta}.}

\item{se.method}{Charactor.
The method for estimating standard error of parameters.
Three options are available.
\describe{
\item{\code{"asymp"}} Obtain estimates of standard error via asymptotic
distribution. This option is applicable to \code{regularize=FALSE} only.
\item{\code{"perturb"}} Obtain estimates of standard error via perturbation method.
This option is applicable to both \code{regularize=FALSE} and \code{regularize=TRUE}.
\item{\code{NULL}} Standard error is not to be estimated.
}}

\item{M}{Number of resampling in perturbation.}

\item{seed_num}{Number of seed to control randomness of perturbation term generated.}

\item{regularize}{If the objective is to fit a regularized linear regression
for variable selection, \code{regularize = TRUE}. By default,
\code{regularize = FALSE}.}

\item{data_pair}{Matrix. The pairwise version of original data obtained from
function \code{pair_trans()$data_pair}.}

\item{newcov_pair}{Matrix. The pairwise transformed covariates obtained from
function \code{pair_trans()$newcov_pair}.}

\item{obs_id}{Vector. The indice of missingness for subjects obtained from
function \code{pair_trans}.}

\item{w_hat}{Matrix. The estimates of \eqn{W} matrix in the offset term
obtained by \code{pair_trans}.}
}
\value{
A vector. The standard error of estimators in unregularized regression.
}
\description{
Estimate standard error using asymptotic formula for unregularized regression.
}
