% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_trans.R
\name{pair_trans}
\alias{pair_trans}
\title{Generate covariates to fit the logistic regression,
which is the form of reparameterized objective function.}
\usage{
pair_trans(y, z, u, eta_est, z.sigma_est)
}
\arguments{
\item{y}{Vector. The response variable subject to missing data.}

\item{z}{Vector. The shadow variable (fully observed).}

\item{u}{Matrix. The covariate matrix excluding shadow variable \eqn{z} (fully observed).}

\item{eta_est}{The estimates for coefficients in model \eqn{Z} given\eqn{U}.}

\item{z.sigma_est}{The estimates for \eqn{\sigma} in model \eqn{Z} given\eqn{U}.}
}
\value{
A list:
\describe{
\item{data_pair}{The matrix of pairwised \eqn{y}, \eqn{z}, \eqn{u}.}
\item{newcov_pair}{The matrix of pairwised transformed variables to be fit in the objective function.}
\item{obs_id}{The indice of missingness for subjects in \code{data}.}
\item{w_hat}{The estimates of \eqn{W} matrix in the offset term.}
}
}
\description{
\code{pair_trans} is used to generate transformed covariates in pairwise form.
}
