% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbate.R
\name{perturbate}
\alias{perturbate}
\title{Obtain the estimates for standard error via perturbation.}
\usage{
perturbate(M, y, seed_num, regularize = FALSE, data_pair, newcov_pair, obs_id)
}
\arguments{
\item{M}{Number of resampling in perturbation.}

\item{y}{Vector. The outcome variable.}

\item{seed_num}{Number of seed to control randomness of perturbation term generated.}

\item{regularize}{If the objective is to fit a regularized linear regression
for variable selection, \code{regularize = TRUE}. By default, \code{regularize = FALSE}.}

\item{data_pair}{Matrix. The pairwise version of original data obtained from
function \code{pair_trans()$data_pair}.}

\item{newcov_pair}{Matrix. The pairwise transformed covariates obtained from
function \code{pair_trans()$newcov_pair}.}

\item{obs_id}{Vector. The indice of missingness for subjects obtained from
function \code{pair_trans()$obs_id}.}
}
\value{
A matrix. \code{M} sets of estimates based on perturbated objective function.
}
\description{
Obtain the estimates for standard error via perturbation.
}
