% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_asymp.R
\name{se_asymp}
\alias{se_asymp}
\title{Generate the covariance matrix of estimators in unregularized regression.}
\usage{
se_asymp(y, z, u, gamma_tilde_hat, beta_hat, eta_est, z.sigma, A_est, B_est,
  G_est)
}
\arguments{
\item{y}{Vector. The outcome variable}

\item{z}{Vector. The shadow variable (fully observed).}

\item{u}{Matrix. The covariate matrix \eqn{u}.}

\item{gamma_tilde_hat}{Initial estimates of scaled \eqn{\gamma}.}

\item{beta_hat}{Initial estimates of \eqn{\beta}.}

\item{z.sigma}{The estimates for \eqn{\sigma} in model \eqn{Z} given\eqn{U}.}

\item{A_est}{Matrix. The estimates of \eqn{A} matrix obtained by \code{AB_matrix}.}

\item{B_est}{Matrix. The estimates of \eqn{B} matrix obtained by \code{AB_matrix}.}

\item{G_est}{Matrix. The estimates of \eqn{G} matrix obtained by \code{G_matrix}}
}
\value{
A vector. The standard error of estimators in unregularized regression.
}
\description{
Generate the covariance matrix of estimators in unregularized regression.
}
