% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIsMiss.R
\name{SIsMiss}
\alias{SIsMiss}
\title{Fit the primary linear model with Nuissance-Free conditional
likelihood method.}
\usage{
SIsMiss(y, z, u, regularize = FALSE, cov.names = NULL, se.method = NULL,
  CI.alpha = NULL, M = 500, seed_num = 123)
}
\arguments{
\item{y}{Vector. The response variable subject to missing data.}

\item{z}{Vector. The shadow variable (fully observed).}

\item{u}{Matrix. The covariate matrix excluding
shadow variable \eqn{z} (fully observed).}

\item{regularize}{Logical. If \code{regularize=TRUE}, a regularized linear
model is going to be fitted. Adaptive LASSO penalty is adopt. Default
is \code{regularize=FALSE}.}

\item{cov.names}{Vector. The vector of names for all the covariates.
The first element is the name of variable \eqn{z},
followed by the name of variable \eqn{u}.}

\item{se.method}{Charactor.
The method for estimating standard error of parameters.
Three options are available.
\describe{
\item{\code{"asymp"}} Obtain estimates of standard error via asymptotic
distribution. This option is applicable to \code{regularize=FALSE} only.
\item{\code{"perturb"}} Obtain estimates of standard error via
perturbation method. This option is applicable to both
\code{regularize=FALSE} and \code{regularize=TRUE}.
\item{\code{NULL}} Standard error is not to be estimated.
}}

\item{CI.alpha}{\eqn{\alpha} value for confidence interval of each parameter.
\code{CI.alpha} must be a value between (0, 1).
If \code{CI.alpha=NULL}, confidence interval will not be returned.}

\item{M}{Number of resampling in perturbation.}

\item{seed_num}{Number of seed to control randomness of perturbation
term generated.}
}
\value{
A table. Summary of estimates and standard error presented in the table
with \eqn{p+1} rows and 2 columns.
}
\description{
Fit the primary linear model with Nuissance-Free conditional
likelihood method.
}
